Move settings.view.ts
