+++ app/src/gui/home-tab/views/settings/settings.view.ts
